Extract RetryConfig and RetryPolicy to executor/retry.rs module
First small step in executor.rs refactoring. This change moves only
the retry-related types (RetryConfig and RetryPolicy) to a dedicated
module while maintaining all existing APIs and functionality.

This focused approach avoids the complexity of large refactoring
attempts and maintains clean, testable code throughout the process.
All tests pass with no behavioral changes.
